+++ packages/engine/Source/DataSources/DataSourceDisplay.js
@@ -330,7 +330,10 @@ DataSourceDisplay.prototype.update = function (time) {
   if (!this._ready && result) {
     this._scene.requestRender();
   }
-  this._ready = result;
+
+  // once the DataSourceDisplay is ready it should stay ready to prevent
+  // entities from breaking updates when they become "un-ready"
+  this._ready = this._ready || result;
 
   return result;
 };
@@ -386,7 +389,7 @@ DataSourceDisplay.prototype.getBoundingSphere = function (
   Check.defined("result", result);
   //>>includeEnd('debug');
 
-  if (!this._ready && !allowPartial) {
+  if (!this._ready) {
     return BoundingSphereState.PENDING;
   }